pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = ""
authors = [
    {name = "AnnN",email = "anechytailenko@kse.org.ua"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib (>=3.10.8,<4.0.0)",
    "memory-profiler (>=0.61.0,<0.62.0)",
    "polars (>=1.38.1,<2.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "kagglehub (>=1.0.0,<2.0.0)",
    "pyspark (>=4.1.1,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "duckdb (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
